requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
openpyxl==3.1.5
slowapi==0.1.9
sentry-sdk==2.19.0
structlog==25.2.0
email-validator==2.2.0
pydantic-settings==2.6.0
reportlab==4.2.5
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
factory-boy==3.3.1
faker==30.0.0
slowapi==0.1.9